Corrigir tipo de `fk_id_empresa` na migração de `tb_empresa_configuracoes` para `integer`, compatível com `tb_empresas`.

+++ database/migrations/2026_02_19_153241_criar_tabela_empresa_permissoes_processos.php
@@ -15,7 +15,7 @@ return new class extends Migration
         if (!Schema::hasTable('tb_empresa_configuracoes')) {
             Schema::create('tb_empresa_configuracoes', function (Blueprint $table) {
                 $table->id('id_empresa_configuracao');
-                $table->unsignedBigInteger('fk_id_empresa');
+                $table->integer('fk_id_empresa');
                 $table->string('chave', 255); // Ex: 'processos_empresa_pode_ver_inscritos'
                 $table->text('valor')->nullable();
                 $table->string('descricao', 255)->nullable();
